+++ hack/make/.integration-test-helpers
@@ -17,7 +17,7 @@ source "$MAKEDIR/.go-autogen"
 
 integration_api_dirs=${TEST_INTEGRATION_DIR:-"$(
 	find ./integration -type d |
-	grep -vE '(^./integration($|/util)|/testdata)')"}
+	grep -vE '(^./integration($|/internal)|/testdata)')"}
 
 run_test_integration() {
 	[[ "$TESTFLAGS" != *-check.f* ]] && run_test_integration_suites
