+++ integration/system/info_linux_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	req "github.com/docker/docker/integration-cli/request"
-	"github.com/docker/docker/integration/util/request"
+	"github.com/docker/docker/integration/internal/request"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/net/context"
