+++ integration/system/version_test.go
@@ -3,7 +3,7 @@ package system // import "github.com/docker/docker/integration/system"
 import (
 	"testing"
 
-	"github.com/docker/docker/integration/util/request"
+	"github.com/docker/docker/integration/internal/request"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/net/context"
